requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
faiss-cpu
pypdf
tiktoken
numpy
tenacity
